nextner/requirements.txt
Django==4.1.4
django-nested-admin==4.0.2
django-nested-formset==0.1.4
